+++ src/main/java/christmas/domain/type/Badge.java
@@ -1,4 +1,4 @@
-package christmas.domian.type;
+package christmas.domain.type;
 
 public enum Badge {
     STAR("별", 5_000),
