+++ src/test/java/christmas/TestConstants.java
@@ -1,6 +1,6 @@
 package christmas;
 
-import christmas.domian.type.Menu;
+import christmas.domain.type.Menu;
 
 import java.util.Map;
