+++ src/test/java/christmas/util/OrderValidatorTest.java
@@ -1,8 +1,7 @@
 package christmas.util;
 
-import christmas.domian.type.Menu;
+import christmas.domain.type.Menu;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
